Move NMS to separate file.
