+++ keras_retinanet/layers/__init__.py
@@ -1,1 +1,2 @@
-from ._misc import NonMaximumSuppression, RegressBoxes, UpsampleLike, Anchors, ClipBoxes
+from ._misc import RegressBoxes, UpsampleLike, Anchors, ClipBoxes
+from .nms import NonMaximumSuppression
